Unify quantity types

+++ src/main/java/com/vpe/finalstore/cart/dtos/CartItemUpdateDto.java
@@ -10,5 +10,5 @@ public class CartItemUpdateDto {
     @NotNull(message = "Quantity is required")
     @Min(value = 1, message = "Quantity must be greater than zero")
     @Max(value = 10_000, message = "Quantity must be less than or equal to 10_000")
-    private short quantity;
+    private Integer quantity;
 }
